+++ packages/message-sdk/src/components/Inbox.tsx
@@ -7,6 +7,7 @@ import type {
 } from "@coral-xyz/common";
 import { BACKEND_API_URL } from "@coral-xyz/common";
 import {
+  BubbleTopLabel,
   EmptyState,
   refreshGroupsAndFriendships,
 } from "@coral-xyz/react-common";
@@ -122,7 +123,7 @@ export function InboxInner() {
           requestCount > 0) && (
           <>
             {searchFilter.length >= 3 && (
-              <div className={classes.topLabel}>Your contacts</div>
+              <BubbleTopLabel text="Your contacts" />
             )}
             <div style={{ paddingBottom: "16px" }}>
               <MessageList
@@ -150,13 +151,13 @@ export function InboxInner() {
           </>
         )}
       {searchFilter.length >= 3 && searchedUsersDistinct.length !== 0 && (
-        <div style={{ marginTop: 30 }}>
-          <div className={classes.topLabel}>Other people</div>
+        <>
+          <BubbleTopLabel text="Other people" />
           <UserList
             users={searchedUsersDistinct}
             setMembers={setSearchResults}
           />
-        </div>
+        </>
       )}
       {allChats &&
         allChats.length !== 0 &&
